+++ pkg/sign/hash.go
